chore(plot): move LineSource class into timeseries_source.py
