+++ src/caret_analyze/plot/visualize_lib/bokeh/bokeh_source.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 from collections import defaultdict
 from logging import getLogger
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 from bokeh.models import GlyphRenderer, HoverTool, Legend
 from bokeh.plotting import ColumnDataSource
@@ -25,7 +25,7 @@ import numpy as np
 
 from ....common import ClockConverter
 from ....exceptions import InvalidArgumentError
-from ....record import Clip, RecordsInterface
+from ....record import Clip
 from ....runtime import (CallbackBase, Communication, Publisher, Subscription,
                          SubscriptionCallback, TimerCallback)
 
@@ -352,111 +352,6 @@ class CallbackSchedBarSource:
         })
 
         return bar_source
-
-
-class LineSource:
-    """Class to generate timeseries line sources."""
-
-    def __init__(
-        self,
-        legend_manager: LegendManager,
-        target_object: TimeSeriesTypes,
-        frame_min,
-        xaxis_type: str,
-    ) -> None:
-        self._legend_keys = LegendKeys('timeseries', target_object)
-        self._hover = HoverCreator(self._legend_keys)
-        self._legend_source = LegendSource(legend_manager, self._legend_keys)
-        self._frame_min = frame_min
-        self._xaxis_type = xaxis_type
-
-    def create_hover(self, options: dict = {}) -> HoverTool:
-        """
-        Create HoverTool based on the legend keys.
-
-        Parameters
-        ----------
-        options : dict, optional
-            Additional options, by default {}
-
-        Returns
-        -------
-        HoverTool
-
-        """
-        return self._hover.create(options)
-
-    def generate(
-        self,
-        target_object: TimeSeriesTypes,
-        timeseries_records: RecordsInterface,
-    ) -> ColumnDataSource:
-        """
-        Generate a line source for timeseries figure.
-
-        Parameters
-        ----------
-        target_object : TimeSeriesTypes
-            TimeSeriesPlotTypes = Union[
-                CallbackBase, Communication, Union[Publisher, Subscription]
-            ]
-        timeseries_records : RecordsInterface
-            Records containing timeseries data.
-
-        Returns
-        -------
-        bokeh.plotting.ColumnDataSource
-            Line source for timeseries figure.
-
-        """
-        line_source = ColumnDataSource(data={
-            k: [] for k in (['x', 'y'] + self._legend_keys.to_list())
-        })
-        legend_source = self._legend_source.generate(target_object)
-        x_item, y_item = self._get_x_y(timeseries_records)
-        for x, y in zip(x_item, y_item):
-            line_source.stream({**{'x': [x], 'y': [y]}, **legend_source})  # type: ignore
-
-        return line_source
-
-    def _get_x_y(
-        self,
-        timeseries_records: RecordsInterface
-    ) -> Tuple[List[Union[int, float]], List[Union[int, float]]]:
-        def ensure_not_none(
-            target_seq: Sequence[Optional[Union[int, float]]]
-        ) -> List[Union[int, float]]:
-            """
-            Ensure the inputted list does not include None.
-
-            Notes
-            -----
-            The timeseries_records is implemented not to include None,
-            so if None is included, an AssertionError is output.
-
-            """
-            not_none_list = [_ for _ in target_seq if _ is not None]
-            assert len(target_seq) == len(not_none_list)
-
-            return not_none_list
-
-        ts_column = timeseries_records.columns[0]
-        value_column = timeseries_records.columns[1]
-        timestamps = ensure_not_none(timeseries_records.get_column_series(ts_column))
-        values = ensure_not_none(timeseries_records.get_column_series(value_column))
-        if 'latency' in value_column.lower() or 'period' in value_column.lower():
-            values = [v*10**(-6) for v in values]  # [ns] -> [ms]
-
-        x_item: List[Union[int, float]]
-        y_item: List[Union[int, float]] = values
-        if self._xaxis_type == 'system_time':
-            x_item = [(ts-self._frame_min)*10**(-9) for ts in timestamps]
-        elif self._xaxis_type == 'index':
-            x_item = list(range(0, len(values)))
-        elif self._xaxis_type == 'sim_time':
-            x_item = timestamps
-
-        return x_item, y_item
 
 
 class LegendManager:
